# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/destinasi": {
      "post": {
        "operationId": "DestinasiController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDestinasiDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Destinasi berhasil dibuat.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDestinasiDto"
                }
              }
            }
          }
        },
        "summary": "Tambah destinasi baru",
        "tags": [
          "destinasi"
        ]
      },
      "get": {
        "operationId": "DestinasiController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Daftar destinasi berhasil diambil.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CreateDestinasiDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Ambil semua destinasi",
        "tags": [
          "destinasi"
        ]
      }
    },
    "/destinasi/{id}": {
      "get": {
        "operationId": "DestinasiController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detail destinasi ditemukan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDestinasiDto"
                }
              }
            }
          },
          "404": {
            "description": "Destinasi tidak ditemukan."
          }
        },
        "summary": "Ambil detail destinasi berdasarkan ID",
        "tags": [
          "destinasi"
        ]
      },
      "patch": {
        "operationId": "DestinasiController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDestinasiDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Destinasi berhasil diperbarui.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDestinasiDto"
                }
              }
            }
          },
          "404": {
            "description": "Destinasi tidak ditemukan."
          }
        },
        "summary": "Perbarui sebagian data destinasi berdasarkan ID",
        "tags": [
          "destinasi"
        ]
      },
      "delete": {
        "operationId": "DestinasiController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Destinasi berhasil dihapus.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDestinasiDto"
                }
              }
            }
          },
          "404": {
            "description": "Destinasi tidak ditemukan."
          }
        },
        "summary": "Hapus destinasi berdasarkan ID",
        "tags": [
          "destinasi"
        ]
      }
    },
    "/destinasi/{id}/ulasan": {
      "get": {
        "operationId": "DestinasiController_findUlasanByDestinasi",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar ulasan berhasil diambil."
          },
          "404": {
            "description": "Destinasi tidak ditemukan."
          }
        },
        "summary": "Ambil semua ulasan berdasarkan ID destinasi",
        "tags": [
          "destinasi"
        ]
      }
    }
  },
  "info": {
    "title": "Wisataku API",
    "description": "Dokumentasi API untuk Manajemen Destinasi Wisata",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "CreateDestinasiDto": {
        "type": "object",
        "properties": {
          "nama": {
            "type": "string",
            "example": "Candi Borobudur"
          },
          "kategori": {
            "type": "string",
            "example": "Budaya"
          },
          "lokasi": {
            "type": "string",
            "example": "Magelang, Jawa Tengah"
          },
          "hargaTiket": {
            "type": "number",
            "example": 50000
          }
        },
        "required": [
          "nama",
          "kategori",
          "hargaTiket"
        ]
      },
      "UpdateDestinasiDto": {
        "type": "object",
        "properties": {
          "nama": {
            "type": "string",
            "example": "Candi Borobudur"
          },
          "kategori": {
            "type": "string",
            "example": "Budaya"
          },
          "lokasi": {
            "type": "string",
            "example": "Magelang, Jawa Tengah"
          },
          "hargaTiket": {
            "type": "number",
            "example": 50000
          }
        }
      }
    }
  }
}